requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dashscope==1.25.2
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.123.10
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
google-auth==2.43.0
google-genai==1.53.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
llvmlite==0.44.0
loguru==0.7.3
markdown==3.10
markupsafe==2.1.5
mecab-python3==1.0.12
mpmath==1.3.0
multidict==6.7.0
networkx==3.3
nltk==3.9.2
numba==0.61.2
numpy==1.26.4
openai==2.9.0
pillow==11.3.0
pip==25.3
pipecat-ai @ git+https://github.com/pipecat-ai/pipecat.git@4cefe1357c0411c262386233b8f5c889bfbbf0f5
propcache==0.4.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyloudnorm==0.1.1
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rsa==4.9.1
scipy==1.15.3
setuptools==80.9.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.50.0
style-bert-vits2==2.7.0
sympy==1.14.0
tenacity==9.1.2
torch==2.9.1+cpu
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.0
unidic-lite==1.0.8
uvicorn==0.38.0
wait-for2==0.4.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
